decoder/vision/track_b/requirements.txt
absl-py==2.3.1
accelerate==1.0.1
aim==3.25.1
aim-ui==3.25.1
aimrecords==0.0.7
aimrocks==0.5.2
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
awscli==1.36.3
base58==2.0.1
black==23.12.1
blessed==1.25.0
boto3==1.35.62
botocore==1.35.62
braceexpand==0.1.7
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decopatch==1.4.10
deepspeed==0.16.1
diffusers==0.32.2
docutils==0.16
einops==0.8.1
fastapi==0.121.3
filelock==3.20.0
fire==0.7.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.7.1
gpustat==1.1.1
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
hjson==3.1.0
huggingface-hub==0.36.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
isort==5.13.2
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
liger_kernel==0.6.4
lightning==2.5.0
lightning-utilities==0.15.2
makefun==1.16.0
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
numpy==2.3.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.2
pathspec==0.12.1
peft==0.18.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
pytest-cases==3.9.1
pytest-datadir==1.8.0
pytest-regressions==2.8.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
RestrictedPython==8.1
rich==14.0.0
rsa==4.7.2
s3transfer==0.10.4
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.47.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
timm==1.0.22
tokenizers==0.22.1
torch==2.6.0
torch-fidelity==0.3.0
torchmetrics==1.4.3
torchvision==0.21.0
tqdm==4.67.1
transformers==4.57.0
triton==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wandb==0.23.1
wcwidth==0.2.14
webdataset==0.2.111
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.22.0
zipp==3.23.0
